docshell/test/iframesandbox/test_parent_navigation_by_location.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=785310
html5 sandboxed iframe should not be able to perform top navigation with scripts allowed
-->
<head>
<meta charset="utf-8">
<title>Test for Bug 785310 - iframe sandbox parent navigation by location tests</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

<script>
  SimpleTest.waitForExplicitFinish();

  function runScriptNavigationTest(testCase) {
    window.onmessage = function(event) {
      if (event.data != "parentIframe") {
        ok(false, "event.data: got '" + event.data + "', expected 'parentIframe'");
      }
      ok(false, testCase.desc + " - parent navigation was NOT blocked");
      runNextTest();
    };
    try {
      window.parentIframe.childIframe.eval(testCase.script);
    } catch (e) {
      ok(true, testCase.desc + " - " + e.message);
      runNextTest();
    }
  }

  var testCaseIndex = -1;
  var testCases = [
    {
      desc: "Test 1: parent.location.replace should be blocked even when sandboxed with allow-same-origin allow-top-navigation",
      script: "parent.location.replace('file_parent_navigation_by_location.html')",
    },
    {
      desc: "Test 2: parent.location.assign should be blocked even when sandboxed with allow-same-origin allow-top-navigation",
      script: "parent.location.assign('file_parent_navigation_by_location.html')",
    },
    {
      desc: "Test 3: parent.location.href should be blocked even when sandboxed with allow-same-origin allow-top-navigation",
      script: "parent.location.href = 'file_parent_navigation_by_location.html'",
    },
    {
      desc: "Test 4: parent.location.hash should be blocked even when sandboxed with allow-same-origin allow-top-navigation",
      script: "parent.location.hash = 'wibble'",
    },
  ];

  function runNextTest() {
    ++testCaseIndex;
    if (testCaseIndex == testCases.length) {
      SimpleTest.finish();
      return;
    }

    runScriptNavigationTest(testCases[testCaseIndex]);
  }

  window.onmessage = runNextTest;
</script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=785310">Mozilla Bug 785310</a>
<p id="display"></p>
<div id="content">
Tests for Bug 785310
</div>

<iframe name="parentIframe" src="file_parent_navigation_by_location.html"></iframe>

</body>
</html>
